add AutoServices

+++ app/components/auto_services/card.rb
@@ -19,31 +19,47 @@ class Components::AutoServices::Card < Components::Base
       p(class: "opacity-80") { @service.notes } if @service.notes.present?
 
       if authenticated? && @entry.user_id == Current.user.id && request.params[:mode] == "services"
-        div(class: "flex flex-wrap gap-2 pt-1") do
-          a(href: edit_auto_service_path(@entry), class: "btn btn-xs") { "Изменить" }
-          a(
-            href: auto_service_path(@entry),
-            class: "btn btn-xs btn-outline",
-            data: { turbo_method: :delete, turbo_confirm: "Удалить услугу в корзину?" }
-          ) { "Удалить" }
+        div(class: "flex flex-col") do
+          div(class: "flex flex-wrap gap-2 py-2") do
+            unless @entry.trash?
+              a(href: edit_auto_service_path(@entry, mode: "services"),
+                class: "btn btn-xs btn-outline btn-info"
+              ) { lucide_icon("pencil", size: 14) }
+            end
 
-          a(
-            href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_MANUAL),
-            class: "btn btn-xs #{@service.manual? ? 'btn-success' : 'btn-ghost'}",
-            data: { turbo_method: :patch }
-          ) { "Активен" }
+            if @entry.trash?
+              a(href: restore_auto_service_path(@entry, mode: "services"),
+                class: "btn btn-xs btn-success",
+                data: { turbo_method: :patch, turbo_confirm: "Восстановить услугу?" }
+              ) { "Восстановить" }
+            else
+              a(href: auto_service_path(@entry, mode: "services"),
+                class: "btn btn-xs btn-outline btn-error",
+                data: { turbo_method: :delete, turbo_confirm: "Удалить услугу в корзину?" }
+              ) { lucide_icon("trash", size: 14) }
+            end
+          end
 
-          a(
-            href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_SCHEDULE),
-            class: "btn btn-xs #{@service.schedule? ? 'btn-success' : 'btn-ghost'}",
-            data: { turbo_method: :patch }
-          ) { "Активность по графику" }
+          div(class: "flex flex-wrap gap-2 py-2 min-w-[350px]") do
+            unless @entry.trash?
+              a(href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_MANUAL, mode: "services"),
+                class: "btn btn-xs #{@service.manual? ? 'btn-success' : 'btn-ghost'}",
+                data: { turbo_method: :patch }
+              ) { "Активен" }
 
-          a(
-            href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_OFF),
-            class: "btn btn-xs #{@service.off? ? 'btn-error' : 'btn-ghost'}",
-            data: { turbo_method: :patch }
-          ) { "Выключен" }
+              a(
+                href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_SCHEDULE, mode: "services"),
+                class: "btn btn-xs #{@service.schedule? ? 'btn-success' : 'btn-ghost'}",
+                data: { turbo_method: :patch }
+              ) { "Активность по графику" }
+
+              a(
+                href: set_activity_auto_service_path(@entry, activity_state: AutoService::STATE_OFF, mode: "services"),
+                class: "btn btn-xs #{@service.off? ? 'btn-error' : 'btn-ghost'}",
+                data: { turbo_method: :patch }
+              ) { "Выключен" }
+            end
+          end
         end
       end
     end
